+++ windows/texture_bridge.h
@@ -1,16 +1,13 @@
 #pragma once
 
-#include <flutter/texture_registrar.h>
 #include <windows.graphics.capture.h>
-#include <winrt/Windows.UI.Composition.Desktop.h>
-#include <winrt/Windows.UI.Composition.h>
+
+#include <wrl.h>
 
 #include <cstdint>
+#include <functional>
 
 #include "graphics_context.h"
-#include "util/capture.desktop.interop.h"
-
-using namespace winrt::Windows::Graphics::Capture;
 
 typedef struct {
   size_t width;
@@ -23,13 +20,11 @@ class TextureBridge {
   typedef std::function<void(Size size)> SurfaceSizeChangedCallback;
 
   TextureBridge(GraphicsContext* graphics_context,
-                winrt::Windows::UI::Composition::Visual surface);
-  ~TextureBridge();
+                ABI::Windows::UI::Composition::IVisual* visual);
+  virtual ~TextureBridge();
 
   bool Start();
   void Stop();
-
-  const FlutterDesktopPixelBuffer* CopyPixelBuffer(size_t width, size_t height);
 
   void SetOnFrameAvailable(FrameAvailableCallback callback) {
     frame_available_ = std::move(callback);
@@ -41,37 +36,29 @@ class TextureBridge {
 
   void NotifySurfaceSizeChanged();
 
- private:
+ protected:
   bool is_running_ = false;
 
   const GraphicsContext* graphics_context_;
 
   FrameAvailableCallback frame_available_;
   SurfaceSizeChangedCallback surface_size_changed_;
-
-  Size staging_texture_size_ = {0, 0};
   std::atomic<bool> needs_update_ = false;
+  winrt::com_ptr<ID3D11Texture2D> last_frame_;
 
-  winrt::Windows::Graphics::Capture::GraphicsCaptureItem capture_item_{nullptr};
-  winrt::Windows::Graphics::Capture::Direct3D11CaptureFramePool frame_pool_{
-      nullptr};
+  winrt::com_ptr<ABI::Windows::Graphics::Capture::IGraphicsCaptureItem>
+      capture_item_;
+  winrt::com_ptr<ABI::Windows::Graphics::Capture::IDirect3D11CaptureFramePool>
+      frame_pool_;
+  winrt::com_ptr<ABI::Windows::Graphics::Capture::IGraphicsCaptureSession>
+      capture_session_;
 
-  winrt::Windows::Graphics::Capture::GraphicsCaptureSession capture_session_{
-      nullptr};
-  winrt::Windows::Graphics::Capture::Direct3D11CaptureFramePool::
-      FrameArrived_revoker frame_arrived_;
+  EventRegistrationToken on_closed_token_ = {};
+  EventRegistrationToken on_frame_arrived_token_ = {};
 
-  winrt::com_ptr<ID3D11Texture2D> staging_texture_{nullptr};
+  void OnFrameArrived();
 
-  std::unique_ptr<uint8_t> backing_pixel_buffer_;
-  std::unique_ptr<FlutterDesktopPixelBuffer> pixel_buffer_;
-
-  void OnFrameArrived(
-      winrt::Windows::Graphics::Capture::Direct3D11CaptureFramePool const&
-          sender,
-      winrt::Windows::Foundation::IInspectable const&);
-
-  void CopyFrame(ID3D11Texture2D* src_texture);
-  void EnsureStagingTexture(uint32_t width, uint32_t height,
-                            bool& is_exact_size);
+  // corresponds to DXGI_FORMAT_B8G8R8A8_UNORM
+  static constexpr auto kPixelFormat = ABI::Windows::Graphics::DirectX::
+      DirectXPixelFormat::DirectXPixelFormat_B8G8R8A8UIntNormalized;
 };